test: split the facts-stream suite and prune impossible inputs
The facts-stream tests were one 1770-line module; they split along their
own section banners, and the document factory six subtrees were reaching
into cli/stats' conftest for moves to tests/spindoctor/conftest.py.

The pruned tests all needed a navigation document carrying a malformed
field. Only the pipeline writes navigation documents, and its own types
settle every field: a success result cannot lack an offset, a technique
covariance is validated square and positive-semidefinite, every matrix is
validated before serialization, and both frame identities are int-typed.
Tests for files that are not navigation documents at all stay, since a
results tree really does hold those.
