+++ tests/spindoctor/cli/stats/test_ingest_cloud_tasks_tallies.py
@@ -18,11 +18,20 @@ from typing import Any
 
 import pdslogger
 import pytest
+from tests.spindoctor.conftest import (
+    index_url,
+)
 
 from spindoctor.cli.stats.ingest.tasks import _LARGEST_RUN_ROW_COUNT, _share_tally
 from spindoctor.results_index import normalize_root_url
 
-from .conftest import build_tree, complete, fan_out, index_url, reported, run_rows
+from .conftest import (
+    build_tree,
+    complete,
+    fan_out,
+    reported,
+    run_rows,
+)
 
 # ---------------------------------------------------------------------------
 # The shape of a tally
