+++ tests/spindoctor/dataset/test_results_filter_index.py
@@ -26,7 +26,7 @@ import pytest
 import sqlalchemy
 from filecache import FCPath
 from tests.conftest import child_interpreter_environment
-from tests.spindoctor.cli.stats.conftest import (
+from tests.spindoctor.conftest import (
     index_url,
     ingest_tree,
     metadata_document,
