+++ tests/spindoctor/dataset/test_results_filter_index_divergence.py
@@ -21,7 +21,7 @@ from typing import Any
 
 import pytest
 from filecache import FCPath
-from tests.spindoctor.cli.stats.conftest import (
+from tests.spindoctor.conftest import (
     index_url,
     ingest_tree,
     metadata_document,
