+++ tests/spindoctor/results_index/test_facts_stream_values.py
@@ -1,0 +1,144 @@
+"""The values the agreement of the two storages is made of.
+
+An agreement is worth what the documents behind it carry, so the values a
+comparison could pass while losing are read out one at a time: a covariance
+matrix no pair of per-axis sigmas states, a list whose order is not its sorted
+order, the pair of columns that tells an epoch read out of a dataset index from
+one a navigation run recorded, and an outcome named against one that is absent.
+
+Every read a query blind to the root could answer out of the wrong row is made
+against each of the two roots in turn, which record different values under the
+same stubs.
+"""
+
+import pytest
+from tests.spindoctor.results_index.conftest import (
+    BOTH_ROOTS,
+    ERROR_STUB,
+    FIRST_VALUES,
+    SECOND_VALUES,
+    SUCCESS_STUB,
+    UNLOADED_STUB,
+    TwoRoots,
+    facts_from_index,
+    facts_of,
+)
+
+from spindoctor.nav_records import Selection
+
+
+@pytest.mark.parametrize('which', BOTH_ROOTS)
+def test_a_twist_covariance_survives_whole(two_roots: TwoRoots, which: str) -> None:
+    """Its rotation row and column carry terms no per-axis sigma states.
+
+    Parameters:
+        two_roots: The two ingested roots and their index.
+        which: The root to read.
+    """
+    values = FIRST_VALUES if which == 'first' else SECOND_VALUES
+    found = facts_of(facts_from_index(two_roots, which, Selection()), SUCCESS_STUB)
+    assert found.image['covariance_px2'] == values.twist_covariance
+
+
+@pytest.mark.parametrize('which', BOTH_ROOTS)
+def test_a_per_technique_covariance_survives_whole(two_roots: TwoRoots, which: str) -> None:
+    """Stored as a matrix, so a reader derives the sigmas rather than the reverse.
+
+    Parameters:
+        two_roots: The two ingested roots and their index.
+        which: The root to read.
+    """
+    values = FIRST_VALUES if which == 'first' else SECOND_VALUES
+    found = facts_of(facts_from_index(two_roots, which, Selection()), SUCCESS_STUB)
+    assert [row['covariance_px2'] for row in found.techniques] == [
+        values.technique_covariance,
+        values.technique_covariance,
+    ]
+
+
+@pytest.mark.parametrize('which', BOTH_ROOTS)
+def test_an_exclusion_list_keeps_the_order_it_was_written_in(
+    two_roots: TwoRoots, which: str
+) -> None:
+    """Which is not sorted order, so a storage that sorted the list is caught.
+
+    Parameters:
+        two_roots: The two ingested roots and their index.
+        which: The root to read.
+    """
+    values = FIRST_VALUES if which == 'first' else SECOND_VALUES
+    found = facts_of(facts_from_index(two_roots, which, Selection()), SUCCESS_STUB)
+    assert found.image['excluded_from_consensus'] == values.excluded
+
+
+def test_the_exclusion_list_under_test_is_not_in_sorted_order() -> None:
+    """Without which the test above would hold whether or not order survived."""
+    assert FIRST_VALUES.excluded != sorted(FIRST_VALUES.excluded)
+
+
+@pytest.mark.parametrize('which', BOTH_ROOTS)
+def test_each_roots_frame_id_comes_back_from_that_root(two_roots: TwoRoots, which: str) -> None:
+    """The two roots name two frames, so a read of one may not answer with the other.
+
+    Parameters:
+        two_roots: The two ingested roots and their index.
+        which: The root to read.
+    """
+    values = FIRST_VALUES if which == 'first' else SECOND_VALUES
+    found = facts_of(facts_from_index(two_roots, which, Selection()), SUCCESS_STUB)
+    assert found.image['camera_frame_id'] == values.camera_frame_id
+
+
+@pytest.mark.parametrize('which', BOTH_ROOTS)
+def test_an_image_that_never_loaded_records_its_epoch_under_the_observation(
+    two_roots: TwoRoots, which: str
+) -> None:
+    """The column that says which of the two epochs a document carried.
+
+    Parameters:
+        two_roots: The two ingested roots and their index.
+        which: The root to read.
+    """
+    values = FIRST_VALUES if which == 'first' else SECOND_VALUES
+    found = facts_of(facts_from_index(two_roots, which, Selection()), UNLOADED_STUB)
+    assert found.image['observation_image_et'] == values.observation_et
+
+
+def test_an_image_that_never_loaded_records_no_provenance_epoch(two_roots: TwoRoots) -> None:
+    """The other half of the pair, which is what makes the two columns worth having.
+
+    Parameters:
+        two_roots: The two ingested roots and their index.
+    """
+    found = facts_of(facts_from_index(two_roots, 'first', Selection()), UNLOADED_STUB)
+    assert found.image['provenance_image_et'] is None
+
+
+def test_a_navigated_image_records_its_epoch_as_provenance(two_roots: TwoRoots) -> None:
+    """Which is the case the pair of columns has to tell apart from the other.
+
+    Parameters:
+        two_roots: The two ingested roots and their index.
+    """
+    found = facts_of(facts_from_index(two_roots, 'first', Selection()), SUCCESS_STUB)
+    assert found.image['provenance_image_et'] == 0.0
+
+
+def test_an_image_naming_no_outcome_carries_no_error(two_roots: TwoRoots) -> None:
+    """An absent outcome and a recorded error are different facts.
+
+    Parameters:
+        two_roots: The two ingested roots and their index.
+    """
+    found = facts_of(facts_from_index(two_roots, 'first', Selection()), UNLOADED_STUB)
+    assert found.image['status_error'] is None
+
+
+def test_a_run_that_ended_in_an_error_names_it(two_roots: TwoRoots) -> None:
+    """The vocabulary an error filter matches verbatim.
+
+    Parameters:
+        two_roots: The two ingested roots and their index.
+    """
+    found = facts_of(facts_from_index(two_roots, 'first', Selection()), ERROR_STUB)
+    assert found.image['status_error'] == 'SPICE(SPKINSUFFDATA)'
