+++ tests/spindoctor/results_index/test_record_source.py
@@ -27,7 +27,7 @@ import pdslogger
 import pytest
 import sqlalchemy
 from sqlalchemy.engine import Engine
-from tests.spindoctor.cli.stats.conftest import (
+from tests.spindoctor.conftest import (
     index_url,
     ingest_tree,
     metadata_document,
